phase-v/services/recurring-task-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
httpx==0.27.2
python-dateutil==2.9.0
pydantic==2.11.7
python-dotenv==1.0.1
